At the start of the document I pre-loaded all of the options.tex files in the default.tex.
However, this was a sadly missed style mistake and assumed that these directories would be
the only copies of options.tex and default.tex present.

Broke that single chain of imports out and made every default.tex file to perform an
initial import of the same relative directory options.tex as its first effective action.

Moved a variable initialization to the top level.